Refactor index.html and add dark mode support
