+++ frontend/index.html
@@ -1,35 +1,51 @@
 <!doctype html>
 <html lang="en">
-    <head>
-        <meta charset="UTF-8" />
-        <meta name="viewport" content="width=device-width, initial-scale=1.0" />
 
-        <meta name="Author" content="" />
-        <meta name="Developer" content="" />
-        <meta name="Owner" content="" />
+<head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
 
-        <title>Elephoria</title>
+    <meta name="Author" content="" />
+    <meta name="Developer" content="" />
+    <meta name="Owner" content="" />
 
-        <!-- Favicon Icons -->
-        <link href="apple-touch-icon.png" rel="apple-touch-icon" sizes="180x180" />
-        <link type="image/png" href="favicon-32x32.png" rel="icon" sizes="32x32" />
-        <link type="image/png" href="favicon-16x16.png" rel="icon" sizes="16x16" />
-        <link href="site.webmanifest" rel="manifest" />
-        <link href="safari-pinned-tab.svg" rel="mask-icon" color="#FFB6B9" />
-        <meta name="msapplication-TileColor" content="#FFB6B9" />
-        <meta name="theme-color" content="#FFB6B9" />
-        <meta name="dark-theme-color" content="#1f2937" />
+    <title>Elephoria</title>
 
-        <!-- Preconnect to Google Fonts -->
-        <link rel="preconnect" href="https://fonts.googleapis.com" />
-        <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
+    <!-- Favicon Icons -->
+    <link href="apple-touch-icon.png" rel="apple-touch-icon" sizes="180x180" />
+    <link type="image/png" href="favicon-32x32.png" rel="icon" sizes="32x32" />
+    <link type="image/png" href="favicon-16x16.png" rel="icon" sizes="16x16" />
+    <link href="site.webmanifest" rel="manifest" />
+    <link href="safari-pinned-tab.svg" rel="mask-icon" color="#FFB6B9" />
+    <meta name="msapplication-TileColor" content="#FFB6B9" />
+    <meta name="theme-color" content="#FFB6B9" />
 
-        <!-- Google Fonts -->
-        <link href="https://fonts.googleapis.com/css2?family=Aladin&family=Indie+Flower&display=swap" rel="stylesheet" />
-    </head>
 
-    <body class="bg-primary">
-        <div id="root"></div>
-        <script type="module" src="/src/main.jsx"></script>
-    </body>
+    <!-- Preconnect to Google Fonts -->
+    <link rel="preconnect" href="https://fonts.googleapis.com" />
+    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
+
+    <!-- Google Fonts -->
+    <link href="https://fonts.googleapis.com/css2?family=Aladin&family=Indie+Flower&display=swap" rel="stylesheet" />
+
+    <!-- Dark Mode Initialization Script -->
+    <script>
+        (function () {
+            const darkTheme = localStorage.getItem('dark');
+            const userPrefersDark = window.matchMedia('(prefers-color-scheme: dark)').matches;
+
+            if (darkTheme === 'true' || (!darkTheme && userPrefersDark)) {
+                document.documentElement.classList.add('dark');
+            } else {
+                document.documentElement.classList.remove('dark');
+            }
+        })();
+    </script>
+</head>
+
+<body class="bg-primary">
+    <div id="root"></div>
+    <script type="module" src="/src/main.jsx"></script>
+</body>
+
 </html>
